requirements.txt
Django==4.2.*
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
gdown==5.2.0
ipython==8.36.0
keras==2.10.0
Keras-Preprocessing==1.1.2
librosa==0.10.2.post1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
numba==0.61.2
numpy==1.26.4
opencv-python==4.11.0.86
pandas==2.2.3
pydub==0.25.1
tensorflow==2.10.1
moviepy==2.2.1